Add template and ajax function
