+++ orders/admin.py
@@ -1,7 +1,7 @@
 """Management of admin panel."""
 
 from django.contrib import admin
-from .models import Order, SmartphoneInOrder
+from .models import Order, SmartphoneInOrder, SmartphoneInBasket
 
 
 class SmartphoneInOrderInline(admin.TabularInline):
@@ -25,6 +25,7 @@ class OrderAdmin (admin.ModelAdmin):
 
 
 admin.site.register(Order, OrderAdmin)
+admin.site.register(SmartphoneInBasket)
 
 
 class SmartphoneInOrderAdmin(admin.ModelAdmin):
